fix serializer for large linearizations
Writing the array length as a byte breaks when the linearization of a class
exceeds 127. Since byte is signed, the length will be seen by the
deserializer as a negative number and will trigger the error path. Storage
is cheap, let's use an int instead for the length.

Also fixes up two other locations in the serializer (ScFunctionElementType
and ScModifiersElementType) where byte was used to serialize array
length.

Fixes http://youtrack.jetbrains.com/issue/SCL-5026

+++ src/org/jetbrains/plugins/scala/lang/psi/stubs/elements/ScTemplateParentsElementType.scala
@@ -21,7 +21,7 @@ abstract class ScTemplateParentsElementType[Func <: ScTemplateParents](debugName
         extends ScStubElementType[ScTemplateParentsStub, ScTemplateParents](debugName) {
   def serialize(stub: ScTemplateParentsStub, dataStream: StubOutputStream) {
     val array = stub.getTemplateParentsTypesTexts
-    dataStream.writeByte(array.length)
+    dataStream.writeInt(array.length)
     for (s <- array) {
       dataStream.writeName(s)
     }
@@ -43,7 +43,7 @@ abstract class ScTemplateParentsElementType[Func <: ScTemplateParents](debugName
   }
 
   def deserializeImpl(dataStream: StubInputStream, parentStub: Any): ScTemplateParentsStub = {
-    val length = dataStream.readByte
+    val length = dataStream.readInt
     if (length >= 0) {
       val res = new Array[StringRef](length)
       for (i <- 0 until length) {
